fix(access): bootstrap initial target policy
